Set tutorials to be shown only if the app is opening the first time

+++ Peppermint/ContactsModel.m
@@ -115,7 +115,7 @@
              }
              
              if(--loadContactsTriggerCount > 0) {
-                 NSLog(@"Load contacts method is recalled %d times during the previous query", loadContactsTriggerCount);
+                 NSLog(@"Load contacts method is recalled %lu times during the previous query", (unsigned long)loadContactsTriggerCount);
                  loadContactsTriggerCount = 0;
                  [self refreshContactList];
              }
